requirements.txt
flask==2.3.2
openai==0.27.8
python-dotenv==1.0.0
pydantic==2.1.1
